requirements.txt
flask
flask-wtf
python-dotenv
pymysql
ipython
werkzeug
Flask-SSLify